move onrunqueries to redux
