+++ public/app/features/alerting/state/reducers.ts
@@ -12,11 +12,11 @@ import {
   NotificationChannelOption,
   NotificationChannelState,
   NotifierDTO,
+  QueryGroupOptions,
 } from 'app/types';
 import store from 'app/core/store';
 import { config } from '@grafana/runtime';
 import { PanelQueryRunner } from '../../query/state/PanelQueryRunner';
-import { QueryGroupOptions } from '../../query/components/QueryGroupOptions';
 
 export const ALERT_DEFINITION_UI_STATE_STORAGE_KEY = 'grafana.alerting.alertDefinition.ui';
 const DEFAULT_ALERT_DEFINITION_UI_STATE: AlertDefinitionUiState = { rightPaneSize: 400, topPaneSize: 0.45 };
