+++ public/app/types/index.ts
@@ -15,6 +15,7 @@ export * from './store';
 export * from './ldap';
 export * from './appEvent';
 export * from './angular';
+export * from './query';
 
 import * as CoreEvents from './events';
 export { CoreEvents };
