+++ public/app/types/query.ts
@@ -1,0 +1,24 @@
+import { DataQuery } from '@grafana/data';
+
+export interface QueryGroupOptions {
+  queries: DataQuery[];
+  dataSource: QueryGroupDataSource;
+  maxDataPoints?: number | null;
+  minInterval?: string | null;
+  cacheTimeout?: string | null;
+  timeRange?: {
+    from?: string | null;
+    shift?: string | null;
+    hide?: boolean;
+  };
+}
+
+export interface ChangeQueryOptions {
+  runQueries?: boolean;
+}
+
+interface QueryGroupDataSource {
+  name?: string | null;
+  uid?: string;
+  default?: boolean;
+}
